chore(server): shrink services after ponytail audit
Delete dead wrappers, share CLI install/probe helpers, and fold single-caller skill usage into SkillsService so the services tree stays leaner without behavior changes.
