+++ apps/server/src/services/agent-browser-service.ts
@@ -2,46 +2,19 @@ import type { AgentBrowserStatusResponse } from "@nakama/core";
 import {
   ensureBunGlobalInstallDirs,
   ensureProcessPath,
-  getToolExecutionEnv,
 } from "../lib/ensure-process-path";
+import {
+  buildGlobalPackageInstallPlan,
+  probeCliVersion,
+  runTimedInstallCommand,
+  summarizeInstallOutput,
+} from "./cli-package-install";
 
 const AGENT_BROWSER_PACKAGE = "agent-browser";
 const AGENT_BROWSER_COMMAND = "agent-browser";
 
-interface AgentBrowserInstallPlan {
-  args: string[];
-  command: string;
-  displayCommand: string;
-}
-
-function detectPackageManager(): "npm" | "bun" {
-  if (Bun.which("npm")) {
-    return "npm";
-  }
-
-  if (Bun.which("bun")) {
-    return "bun";
-  }
-
-  return "npm";
-}
-
-function buildAgentBrowserCliInstallPlan(
-  packageManager: "npm" | "bun" = detectPackageManager()
-): AgentBrowserInstallPlan {
-  if (packageManager === "bun") {
-    return {
-      args: ["install", "-g", "--trust", AGENT_BROWSER_PACKAGE],
-      command: "bun",
-      displayCommand: `bun install -g --trust ${AGENT_BROWSER_PACKAGE}`,
-    };
-  }
-
-  return {
-    args: ["install", "-g", AGENT_BROWSER_PACKAGE],
-    command: "npm",
-    displayCommand: `npm install -g ${AGENT_BROWSER_PACKAGE}`,
-  };
+function buildAgentBrowserCliInstallPlan() {
+  return buildGlobalPackageInstallPlan(AGENT_BROWSER_PACKAGE);
 }
 
 export function getAgentBrowserInstallCommand(): string {
@@ -49,93 +22,10 @@ export function getAgentBrowserInstallCommand(): string {
   return `${cliPlan.displayCommand} && ${AGENT_BROWSER_COMMAND} install`;
 }
 
-function extractVersion(stdout: string, stderr: string): string | null {
-  const output = `${stdout}\n${stderr}`.trim();
-  if (!output) {
-    return null;
-  }
-
-  return output.split(/\r?\n/, 1)[0]?.trim() || null;
-}
-
-/** How long a timed-out child gets to honour SIGTERM before it is killed. */
-const SIGTERM_GRACE_MS = 2000;
-
-async function probeAgentBrowserVersion(): Promise<{
-  installed: boolean;
-  version: string | null;
-  missing: boolean;
-}> {
-  const { spawn } = await import("node:child_process");
-  const timeoutMs = 5000;
-
-  return new Promise((resolve) => {
-    let child: ReturnType<typeof spawn>;
-
-    try {
-      child = spawn(AGENT_BROWSER_COMMAND, ["--version"], {
-        env: getToolExecutionEnv(),
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch {
-      resolve({
-        installed: false,
-        missing: true,
-        version: null,
-      });
-      return;
-    }
-
-    let stdout = "";
-    let stderr = "";
-    let killTimeoutId: ReturnType<typeof setTimeout> | undefined;
-
-    // Settings awaits this probe, so a CLI that never answers --version
-    // would otherwise hold GET /v1/settings/agent-browser open forever.
-    // Resolve on the timer instead of waiting for close, because a child
-    // that ignores SIGTERM never emits one. `missing: false` keeps
-    // getAgentBrowserRuntimeStatus from paying the timeout a second time
-    // on its PATH retry.
-    const timeoutId = setTimeout(() => {
-      child.kill("SIGTERM");
-      // A CLI that traps SIGTERM outlives the probe, and its open stdio
-      // pipes then keep this process alive too. Same escalation as
-      // probeHarnessVersion (#338).
-      killTimeoutId = setTimeout(() => child.kill("SIGKILL"), SIGTERM_GRACE_MS);
-      resolve({ installed: false, missing: false, version: null });
-    }, timeoutMs);
-
-    child.stdout?.on("data", (chunk) => {
-      stdout += chunk.toString();
-    });
-    child.stderr?.on("data", (chunk) => {
-      stderr += chunk.toString();
-    });
-    child.once("error", (error) => {
-      clearTimeout(timeoutId);
-      clearTimeout(killTimeoutId);
-      resolve({
-        installed: false,
-        missing: (error as NodeJS.ErrnoException).code === "ENOENT",
-        version: null,
-      });
-    });
-    child.once("close", (code) => {
-      clearTimeout(timeoutId);
-      clearTimeout(killTimeoutId);
-      resolve({
-        installed: code === 0,
-        missing: false,
-        version: code === 0 ? extractVersion(stdout, stderr) : null,
-      });
-    });
-  });
-}
-
 async function getAgentBrowserRuntimeStatus(): Promise<
   Pick<AgentBrowserStatusResponse, "installed" | "version">
 > {
-  const initial = await probeAgentBrowserVersion();
+  const initial = await probeCliVersion(AGENT_BROWSER_COMMAND);
 
   if (initial.installed || !initial.missing) {
     return {
@@ -145,7 +35,7 @@ async function getAgentBrowserRuntimeStatus(): Promise<
   }
 
   ensureProcessPath();
-  const retried = await probeAgentBrowserVersion();
+  const retried = await probeCliVersion(AGENT_BROWSER_COMMAND);
 
   return {
     installed: retried.installed,
@@ -164,8 +54,8 @@ function toAgentBrowserStatusResponse(
     nextStep: ready ? null : "install",
     ready,
     statusMessage: ready
-      ? "agent-browser is installed and ready."
-      : "Install the agent-browser CLI and Chrome on this machine.",
+      ? null
+      : "agent-browser is not installed. Install it to enable browser automation.",
     version: runtime.version,
   };
 }
@@ -177,148 +67,6 @@ export async function getAgentBrowserStatus(): Promise<AgentBrowserStatusRespons
 
 export interface AgentBrowserInstallProgress {
   message: string;
-}
-
-async function runInstallCommand(
-  plan: AgentBrowserInstallPlan,
-  onProgress?: (message: string) => void
-): Promise<{
-  exitCode: number | null;
-  stdout: string;
-  stderr: string;
-  timedOut: boolean;
-}> {
-  const { spawn } = await import("node:child_process");
-  const timeoutMs = 120_000;
-
-  return new Promise((resolve) => {
-    const child = spawn(plan.command, plan.args, {
-      env: getToolExecutionEnv(),
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-    let stdout = "";
-    let stderr = "";
-    let timedOut = false;
-    let stdoutBuffer = "";
-    let stderrBuffer = "";
-
-    const timeoutId = setTimeout(() => {
-      timedOut = true;
-      child.kill("SIGTERM");
-    }, timeoutMs);
-
-    const emitLine = (prefix: "stdout" | "stderr", line: string) => {
-      onProgress?.(`${prefix}: ${line}`);
-    };
-
-    const flushBuffer = (buffer: string, prefix: "stdout" | "stderr") => {
-      let nextBuffer = buffer;
-
-      while (true) {
-        const newlineIndex = nextBuffer.search(/\r?\n/);
-
-        if (newlineIndex < 0) {
-          break;
-        }
-
-        const newlineLength = nextBuffer[newlineIndex] === "\r" ? 2 : 1;
-        const line = nextBuffer.slice(0, newlineIndex).trim();
-        nextBuffer = nextBuffer.slice(newlineIndex + newlineLength);
-
-        if (line) {
-          emitLine(prefix, line);
-        }
-      }
-
-      return nextBuffer;
-    };
-
-    child.stdout?.on("data", (chunk) => {
-      const text = chunk.toString();
-      stdout += text;
-      stdoutBuffer += text;
-      stdoutBuffer = flushBuffer(stdoutBuffer, "stdout");
-    });
-    child.stderr?.on("data", (chunk) => {
-      const text = chunk.toString();
-      stderr += text;
-      stderrBuffer += text;
-      stderrBuffer = flushBuffer(stderrBuffer, "stderr");
-    });
-
-    child.once("error", (error) => {
-      clearTimeout(timeoutId);
-
-      if (stdoutBuffer.trim()) {
-        emitLine("stdout", stdoutBuffer.trim());
-      }
-      if (stderrBuffer.trim()) {
-        emitLine("stderr", stderrBuffer.trim());
-      }
-
-      resolve({
-        exitCode: null,
-        stderr: `${stderr}\n${String(error)}`.trim(),
-        stdout,
-        timedOut,
-      });
-    });
-
-    child.once("close", (exitCode) => {
-      clearTimeout(timeoutId);
-
-      if (stdoutBuffer.trim()) {
-        emitLine("stdout", stdoutBuffer.trim());
-      }
-      if (stderrBuffer.trim()) {
-        emitLine("stderr", stderrBuffer.trim());
-      }
-
-      resolve({
-        exitCode,
-        stderr: stderr.trim(),
-        stdout: stdout.trim(),
-        timedOut,
-      });
-    });
-  });
-}
-
-function summarizeInstallOutput(output: string): string {
-  const lines = output
-    .split(/\r?\n/)
-    .map((line) => line.trim())
-    .filter(Boolean);
-  const meaningful =
-    lines.find((line) => /^error:/i.test(line)) ??
-    lines.find((line) =>
-      /(?:EACCES|ENOENT|EPERM|failed|permission denied)/i.test(line)
-    ) ??
-    lines.find((line) => !/^bun (?:add|install) v/i.test(line)) ??
-    lines[0] ??
-    output.trim();
-  return meaningful.length > 180
-    ? `${meaningful.slice(0, 177)}...`
-    : meaningful;
-}
-
-async function runAgentBrowserCommand(
-  args: string[],
-  onProgress?: (message: string) => void
-): Promise<{
-  exitCode: number | null;
-  stdout: string;
-  stderr: string;
-  timedOut: boolean;
-}> {
-  return runInstallCommand(
-    {
-      args,
-      command: AGENT_BROWSER_COMMAND,
-      displayCommand: `${AGENT_BROWSER_COMMAND} ${args.join(" ")}`,
-    },
-    onProgress
-  );
 }
 
 export async function installAgentBrowser(
@@ -336,7 +84,7 @@ export async function installAgentBrowser(
   emitProgress("Starting agent-browser install.");
   emitProgress(cliPlan.displayCommand);
 
-  const cliResult = await runInstallCommand(cliPlan, emitProgress);
+  const cliResult = await runTimedInstallCommand(cliPlan, emitProgress);
   const cliOutput = [cliResult.stdout, cliResult.stderr]
     .filter(Boolean)
     .join("\n")
@@ -359,7 +107,14 @@ export async function installAgentBrowser(
   ensureProcessPath();
   emitProgress(`${AGENT_BROWSER_COMMAND} install`);
 
-  const browserResult = await runAgentBrowserCommand(["install"], emitProgress);
+  const browserResult = await runTimedInstallCommand(
+    {
+      args: ["install"],
+      command: AGENT_BROWSER_COMMAND,
+      displayCommand: `${AGENT_BROWSER_COMMAND} install`,
+    },
+    emitProgress
+  );
   const browserOutput = [browserResult.stdout, browserResult.stderr]
     .filter(Boolean)
     .join("\n")
