+++ apps/server/src/services/skills-service.ts
@@ -50,24 +50,16 @@ import type {
   StoredSkillRecord,
   StoredSkillUsageRecord,
 } from "@nakama/db";
-import {
-  type SkillUsageRecordingContext,
-  SkillUsageService,
-} from "./skill-usage-service";
 
-export type { SkillUsageRecordingContext };
+export interface SkillUsageRecordingContext {
+  seenCatalogSkillIds: Set<string>;
+  sessionId: string;
+}
 
 const bundledSkillNames = new Set<string>(BUNDLED_SKILL_NAMES);
 
 export class SkillsService {
-  private readonly skillUsageService: SkillUsageService;
-
-  constructor(
-    private readonly db: DatabaseAdapter,
-    skillUsageService?: SkillUsageService
-  ) {
-    this.skillUsageService = skillUsageService ?? new SkillUsageService(db);
-  }
+  constructor(private readonly db: DatabaseAdapter) {}
 
   async syncDiscoveredSkills(): Promise<SyncSkillsResponse> {
     const discovered = await discoverSkills();
@@ -216,7 +208,7 @@ export class SkillsService {
 
     const profileId = options?.profileId?.trim();
     if (profileId) {
-      await this.skillUsageService.recordPatch(orgId, profileId, synced.id);
+      await this.recordPatch(orgId, profileId, synced.id);
     }
 
     return this.getSkill(synced.id);
@@ -403,7 +395,7 @@ export class SkillsService {
       "patched"
     );
 
-    await this.skillUsageService.recordPatch(orgId, profileId, record.id);
+    await this.recordPatch(orgId, profileId, record.id);
 
     return this.getSkill(record.id);
   }
@@ -469,7 +461,7 @@ export class SkillsService {
       "patched"
     );
 
-    await this.skillUsageService.recordPatch(orgId, profileId, record.id);
+    await this.recordPatch(orgId, profileId, record.id);
 
     return this.getSkill(record.id);
   }
@@ -584,12 +576,7 @@ export class SkillsService {
       )
       .filter((skillId): skillId is string => Boolean(skillId));
 
-    void this.skillUsageService.recordCatalogViews(
-      orgId,
-      profileId,
-      skillIds,
-      usageContext
-    );
+    void this.recordCatalogViews(orgId, profileId, skillIds, usageContext);
 
     return composeSkillsCatalog(assigned);
   }
@@ -624,11 +611,7 @@ export class SkillsService {
         )
         .filter((skillId): skillId is string => Boolean(skillId));
 
-      void this.skillUsageService.recordMatches(
-        orgId,
-        profileId,
-        matchedSkillIds
-      );
+      void this.recordMatches(orgId, profileId, matchedSkillIds);
     }
 
     const prompt = composeMatchedSkillsPrompt(matched, {
@@ -645,7 +628,7 @@ export class SkillsService {
     profileId: string
   ): Promise<SkillSummary[]> {
     const records = await this.db.listSkillsForProfile(profileId);
-    const usage = await this.skillUsageService.listForProfile(profileId);
+    const usage = await this.listSkillUsageForProfile(profileId);
     return toSkillSummaries(records, usage);
   }
 
@@ -662,8 +645,100 @@ export class SkillsService {
     return skills.map((record) => toSkillSummary(record));
   }
 
-  getSkillUsageService(): SkillUsageService {
-    return this.skillUsageService;
+  async recordCatalogViews(
+    orgId: string,
+    profileId: string,
+    skillIds: string[],
+    context?: SkillUsageRecordingContext
+  ): Promise<void> {
+    if (skillIds.length === 0) {
+      return;
+    }
+
+    const now = new Date().toISOString();
+
+    for (const skillId of skillIds) {
+      if (context) {
+        const dedupeKey = `${context.sessionId}:${skillId}`;
+        if (context.seenCatalogSkillIds.has(dedupeKey)) {
+          continue;
+        }
+
+        context.seenCatalogSkillIds.add(dedupeKey);
+      }
+
+      await this.safeIncrementSkillUsage({
+        orgId,
+        profileId,
+        skillId,
+        viewDelta: 1,
+        viewedAt: now,
+      });
+    }
+  }
+
+  async recordMatches(
+    orgId: string,
+    profileId: string,
+    skillIds: string[]
+  ): Promise<void> {
+    if (skillIds.length === 0) {
+      return;
+    }
+
+    const now = new Date().toISOString();
+
+    for (const skillId of skillIds) {
+      await this.safeIncrementSkillUsage({
+        orgId,
+        profileId,
+        skillId,
+        useDelta: 1,
+        usedAt: now,
+      });
+    }
+  }
+
+  async recordPatch(
+    orgId: string,
+    profileId: string,
+    skillId: string
+  ): Promise<void> {
+    const now = new Date().toISOString();
+    await this.safeIncrementSkillUsage({
+      orgId,
+      patchDelta: 1,
+      patchedAt: now,
+      profileId,
+      skillId,
+    });
+  }
+
+  async listSkillUsageForProfile(profileId: string) {
+    return this.db.listSkillUsageForProfile(profileId);
+  }
+
+  private async safeIncrementSkillUsage(input: {
+    orgId: string;
+    profileId: string;
+    skillId: string;
+    viewDelta?: number;
+    useDelta?: number;
+    patchDelta?: number;
+    viewedAt?: string;
+    usedAt?: string;
+    patchedAt?: string;
+  }): Promise<void> {
+    try {
+      const assigned = await this.db.listSkillsForProfile(input.profileId);
+      if (!assigned.some((skill) => skill.id === input.skillId)) {
+        return;
+      }
+
+      await this.db.incrementSkillUsage(input);
+    } catch (error) {
+      console.error("Failed to record skill usage:", error);
+    }
   }
 
   private async getAssignedDiscoveredSkills(
