+++ apps/server/src/services/task-service.ts
@@ -5,14 +5,35 @@ import type {
   TaskStatus,
   UpdateTaskRequest,
 } from "@nakama/core";
-import { createId, NakamaApiError } from "@nakama/core";
+import { createId, NakamaApiError, TASK_STATUSES } from "@nakama/core";
 import { canAccessSuperBotProfile } from "@nakama/core/profiles";
 import type { DatabaseAdapter, StoredTaskRecord } from "@nakama/db";
 import type { TaskRunner } from "./task-runner";
-import { isValidTaskStatus, validateTaskInput } from "./task-validate";
 
 /** Caller role context used to gate access to admin-only profiles (e.g. Super Bot). */
 export type ProfileAccess = Parameters<typeof canAccessSuperBotProfile>[0];
+
+function isValidTaskStatus(value: string): value is TaskStatus {
+  return (TASK_STATUSES as readonly string[]).includes(value);
+}
+
+function validateTaskInput(input: {
+  title: string;
+  prompt: string;
+  status?: TaskStatus;
+}): void {
+  if (!input.title.trim()) {
+    throw new Error("Task title is required.");
+  }
+
+  if (!input.prompt.trim()) {
+    throw new Error("Task prompt is required.");
+  }
+
+  if (input.status && !isValidTaskStatus(input.status)) {
+    throw new Error(`Invalid task status: ${input.status}`);
+  }
+}
 
 export class TaskService {
   private taskRunner: TaskRunner | null = null;
